modularization
prettier main.rs, removed the ugly chunks of code and modularized.
no functional differences
