+++ Cargo.toml
@@ -33,3 +33,9 @@ lazy_static = "1.4.0"
 dashmap = "^5.5.3"
 uuid = { version = "1.7.0", features = ["v4"] }
 duration-str = "0.7.1"
+
+
+[profile.release]
+opt-level = 3
+lto = true
+strip = true
